@@ -3,17 +3,15 @@ using MongoDB.Bson;
 using MongoDB.Bson.Serialization.IdGenerators;
 using MongoDB.Bson.Serialization.Serializers;
 
-namespace Mongo.Context.Example.Maps
+namespace Mongo.Context.Example.Maps;
+public class CustomerMap : MongoClassMap<Entities.Customer>
 {
-    public class CustomerMap : MongoClassMap<Entities.Customer>
+    public CustomerMap()
     {
-        public CustomerMap()
-        {
-            AutoMap();
-            SetIgnoreExtraElements(true);
-            MapIdProperty(x => x.Id)
-                .SetSerializer(new StringSerializer(BsonType.ObjectId))
-                .SetIdGenerator(StringObjectIdGenerator.Instance);
-        }
+        AutoMap();
+        SetIgnoreExtraElements(true);
+        MapIdProperty(x => x.Id)
+            .SetSerializer(new StringSerializer(BsonType.ObjectId))
+            .SetIdGenerator(StringObjectIdGenerator.Instance);
     }
 }
