@@ -1,17 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
+﻿namespace Mongo.Context.Example.Entities;
 
-namespace Mongo.Context.Example.Entities
+public class Contact
 {
-    public class Contact
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-        public string Email { get; set; }
-        public Gender Gender { get; set; }
-        public string Phone { get; set; }
-        public bool IsDeleted { get; set; }
+    public string Id { get; set; }
+    public string Name { get; set; }
+    public string Email { get; set; }
+    public Gender Gender { get; set; }
+    public string Phone { get; set; }
+    public bool IsDeleted { get; set; }
 
-        public Address Address { get; set; }
-    }
+    public Address Address { get; set; }
 }
+
